% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_lp_data}
\alias{example_lp_data}
\title{Simulated low-pass WGS dataset for MLLPT}
\format{
A list containing:
\describe{
  \item{tree}{Phylogentic tree constructed from deep WGS data.}
  \item{phydata}{PhyDat object used for the tree construction (includes id attribute).}
  \item{samples}{Low-pass WGS data of two samples (EPICC_C518_C1_G1_L1 and EPICC_C518_D1_G1_L1).}
}
}
\usage{
example_lp_data
}
\description{
A list containing sample tree data (phylo & phyDat object) constructed
from deep WGS samples and a list of two low-pass WGS samples.
}
\keyword{datasets}
