% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_mutation_ids_per_edge}
\alias{get_mutation_ids_per_edge}
\title{Internal function used to get mutation ids for each edge of the tree.}
\usage{
get_mutation_ids_per_edge(tree, phydata)
}
\arguments{
\item{tree}{A tree object (of class phylo).}

\item{phydata}{Phylogenetic data used for the tree construction (of class phyDat).}
}
\value{
A list of mutations gained and loosed on each edge.
}
\description{
Internal function used to get mutation ids for each edge of the tree.
}
