% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{load_genotyping_file}
\alias{load_genotyping_file}
\title{Function to load genotyping data from a file}
\usage{
load_genotyping_file(file, use = NULL, cn_data = NULL)
}
\arguments{
\item{file}{Input file}

\item{use}{Optional vector of mutation ids to use.}

\item{cn_data}{Either a Genomic Ranges object containing cn and mm metadata columns or a data.frame with a id, cn and mm column.}
}
\value{
A tibble containing the genotyping data.
}
\description{
Function to load genotyping data from a file
}
