% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_lp_loglik_edge}
\alias{plot_lp_loglik_edge}
\title{Plot showing log-likelihood of fraction mutated from most likely edge.}
\usage{
plot_lp_loglik_edge(d, labeller_function = NULL, subset = NULL)
}
\arguments{
\item{d}{Object returned from 'add_lowpass_sampled'.}

\item{labeller_function}{(optional) function returning a modified version of sample labels.}

\item{subset}{(optional) character vector defining a subset of samples to plot.}
}
\value{
A ggplot object
}
\description{
Plot showing log-likelihood of fraction mutated from most likely edge.
}
