% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_sample_data}
\alias{plot_sample_data}
\title{Plot showing ML estimate of sample parameters.}
\usage{
plot_sample_data(
  d,
  labeller_function = NULL,
  external_purity_estimate = NULL,
  label_external_purity = "Independent estimate",
  subset = NULL
)
}
\arguments{
\item{d}{Object returned from 'add_lowpass_sampled'.}

\item{labeller_function}{(optional) function returning a modified version of sample labels.}

\item{external_purity_estimate}{(optional) numeric vector of external purity estimates of each sample.}

\item{label_external_purity}{(optional) label to be used in the legend of the external purity estimates.}

\item{subset}{(optional) character vector defining a subset of samples to plot.}
}
\value{
A ggplot object
}
\description{
Plot showing ML estimate of sample parameters.
}
