% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_tree}
\alias{plot_tree}
\title{A internal function used for the plotting of trees}
\usage{
plot_tree(
  tree,
  HI = NULL,
  color_by = NULL,
  linewidth = 1,
  pointsize = 1,
  textsize = 2.5,
  labeller_function = NULL
)
}
\arguments{
\item{tree}{object of class phylo to plot.}

\item{HI}{(optional) HI of the tree (unused).}

\item{color_by}{(optional) lookup table with color labels (NULL).}

\item{linewidth}{(optional) numeric value defining linewidth of tree edges (1).}

\item{pointsize}{(optional)  numeric value defining point size of tree tips  (1).}

\item{textsize}{(optional)  numeric value defining linewidth of tree edges (2.5).}

\item{labeller_function}{(optional) HI of the tree (unused).}
}
\value{
ggplot object
}
\description{
A internal function used for the plotting of trees
}
\examples{
plot_tree(ape::rtree(5))
}
