% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_tree_and_cn_data}
\alias{plot_tree_and_cn_data}
\title{Plot showing tree and CNA data site-by-site.}
\usage{
plot_tree_and_cn_data(
  x,
  cn_data,
  title = "",
  dropped_samples = NULL,
  plot_tree_function = NULL,
  labeller_function = function(x) x,
  ...
)
}
\arguments{
\item{x}{Object returned from 'add_lowpass_sampled' or a object of type 'phylo'.}

\item{cn_data}{CNA data as a named list of genomic ranges objects with a 'cn' metadata column.}

\item{title}{(optional) title for the plot.}

\item{dropped_samples}{(optional) vector of dropped samples.}

\item{plot_tree_function}{(optional) function to plot the tree object.}

\item{labeller_function}{(optional) function returning a modified version of sample labels.}

\item{...}{(optional) arguments passed to the 'plot_tree_function' function.}
}
\value{
A grob.
}
\description{
Plot showing tree and CNA data site-by-site.
}
