% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{remove_root_tip}
\alias{remove_root_tip}
\title{Internal function that removes the reference tip from a tree}
\usage{
remove_root_tip(tree, out)
}
\arguments{
\item{tree}{object of class phylo.}

\item{out}{character vector with the tip name.}
}
\value{
modified tree
}
\description{
Internal function that removes the reference tip from a tree
}
