% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{set_lp_tiplength}
\alias{set_lp_tiplength}
\title{Scale edge length of added LP samples for plotting}
\usage{
set_lp_tiplength(x, frac_height = 0.01)
}
\arguments{
\item{x}{A phylo object.}

\item{frac_height}{The edge length relative to the tree height.}
}
\value{
A phylo object.
}
\description{
Scale edge length of added LP samples for plotting
}
